Refactor(BridegeGame, OutputView, Result): BridgeGame의 결과처리를 Result클래스에 위임하고, 출력할 수 있도록 변경
